pyproject.toml
[tool.poetry]
name = "pttautosign"
version = "1.1.0"
description = "PTT auto sign-in tool with Telegram notifications"
authors = ["crazycat836 <crazycat836@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pyptt = "1.2.0"
python-dotenv = "1.0.1"
telnetlib3 = "2.0.4"
requests = "2.32.3"
websockets = "15.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
pytest-cov = "4.1.0"
black = "23.7.0"
isort = "5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--no-bytecode --cov=. --cov-report=term-missing"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100
